requirements.txt
flask
streamlit
sentence-transformers
transformers
faiss-cpu
